requirements.txt
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
h11==0.12.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pypika-tortoise==0.1.2
python-dotenv==0.19.2
python-http-client==3.3.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rsa==4.8
sendgrid==6.9.3
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.0.3
starlette==0.16.0
tortoise-orm==0.18.0
typing_extensions==4.0.1
uvicorn==0.16.0
